+++ AmazoniamWillowHotel/AmazoniamWillowHotel/Controllers/FacilitieController.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Data.Entity;
+using System.Linq;
+using System.Web;
+using System.Web.Mvc;
+
+namespace AmazoniamWillowHotel.Controllers
+{
+    public class FacilitieController : Controller 
+    {
+        public ActionResult Facilities()
+        {
+            using (var mo = new Models.Hotel_Amazonian_WillowEntities())
+            {
+                ViewData["Facilities"] = mo.Pagina.Where(x => x.nombre == "Facilidades").Include(p => p.Info).Include(p => p.Info.Select(x => x.Imagen1)).ToList();
+            }
+            return View();
+        }
+
+
+
+    }
+}
