+++ AmazoniamWillowHotel/AmazoniamWillowHotel/Controllers/HomeController.cs
@@ -42,23 +42,7 @@ namespace AmazoniamWillowHotel.Controllers
             return View();
         }
 
-        public ActionResult Facilities()
-        {
-            using (var mo = new Models.Hotel_Amazonian_WillowEntities()) {
-                ViewData["Facilities"] = mo.Pagina.Where(x => x.nombre == "Facilidades").Include(p => p.Info).Include(p => p.Info.Select(x => x.Imagen1)).ToList();
-            }
-            return View();
-        }
-
-        public ActionResult InsertFacilitie()
-        {
-            using (var mo = new Models.Hotel_Amazonian_WillowEntities())
-            {
-                ViewData["Status"] = new SelectList(mo.Estado.ToList(), "Id_Estado", "Nombre"); 
-            }
-            //ViewData["Status"] = new SelectList(model.getStatus(), "Id_Estado", "Nombre"); ;
-            return View();
-        }
+      
 
         public ActionResult HowToGet()
         {
