+++ AmazoniamWillowHotel/AmazoniamWillowHotel/Views/Admin/seeAvailableDay.cshtml
@@ -1,0 +1,7 @@
+﻿
+@{
+    ViewBag.Title = "Habitaciones día de hoy";
+}
+
+<h2>seeAvailableDay</h2>
+
