+++ AmazoniamWillowHotel/AmazoniamWillowHotel/Views/Facilitie/Facilities.cshtml
@@ -1,0 +1,55 @@
+﻿@{
+    ViewBag.Title = "Facilidades";
+}
+
+<h2>Facilidades</h2>
+<br />
+<br />
+
+@foreach (var item in (List<AmazoniamWillowHotel.Models.Pagina>)ViewData["Facilities"])
+{
+
+    foreach (var info in (List<AmazoniamWillowHotel.Models.Info>)@item.Info.ToList())
+    {
+        if (info.descripcion != null)
+        {
+            <h3>@info.titulo</h3>
+            <br />
+            var base64ImagenHotel = Convert.ToBase64String(info.Imagen1.imagen1);
+            string imgSrcImagenHotel = String.Format("data:image/gif;base64,{0}", base64ImagenHotel);
+            <div class="col-sm-4">
+                <img class="mostrarModal" width="100%" alt="@info.descripcion" src="@imgSrcImagenHotel">
+            </div>
+        }
+    }
+}
+
+<div class="modal fade" id="myModal" role="dialog">
+    <div class="modal-dialog">
+
+        <div class="modal-content">
+            <div class="modal-header">
+                <button type="button" class="close" data-dismiss="modal">&times;</button>
+                <h4 class="modal-title"></h4>
+            </div>
+            <div class="modal-body">
+            </div>
+        </div>
+
+    </div>
+</div>
+<script>
+    $(document).ready(function () {
+        $(".mostrarModal").click(function () {
+
+            var modal = $("#myModal");
+            var titulo = $(this).attr('alt');
+            var img = $(this).attr('src');
+
+            modal.find('.modal-title').text(titulo);
+            modal.find('.modal-body').html('<img width="100%" src="' + img + '" />');
+            modal.modal();
+
+        });
+    });
+</script>
