--- AmazoniamWillowHotel/AmazoniamWillowHotel/Views/Home/Facilities.cshtml
@@ -1,55 +1,0 @@
-﻿@model List<AmazoniamWillowHotel.Models.Facilidad>
-@{
-    ViewBag.Title = "Facilidades";
-}
-
-<h2>Facilidades</h2>
-<br />
-<br />
-<div class="modal fade" id="myModal" role="dialog">
-    <div class="modal-dialog">
-
-        <div class="modal-content">
-            <div class="modal-header">
-                <button type="button" class="close" data-dismiss="modal">&times;</button>
-                <h4 class="modal-title"></h4>
-            </div>
-            <div class="modal-body">
-            </div>
-        </div>
-
-    </div>
-</div>
-
-<div class="col-sm-12">
-    <div class="row">
-        <div class="col-sm-8">
-
-            <div class="row">
-
-                @foreach (var item in (List<AmazoniamWillowHotel.Models.Facilidad>)ViewData["Facilities"])
-                 {
-                    <div class="col-xs-3">
-                        <img class="mostrarModal" width="120%" alt="@item.Descripcion" src="~/Public/images/@item.Imagen">
-                    </div>
-                  }
-            </div>
-            <br />
-        </div>
-    </div>
-</div>
-    <script>
-        $(document).ready(function () {
-            $(".mostrarModal").click(function () {
-
-                var modal = $("#myModal");
-                var titulo = $(this).attr('alt');
-                var img = $(this).attr('src');
-
-                modal.find('.modal-title').text(titulo);
-                modal.find('.modal-body').html('<img width="100%" src="' + img + '" />');
-                modal.modal();
-
-            });
-        });
-    </script>
