+++ AmazoniamWillowHotel/AmazoniamWillowHotel/Views/Shared/_Layout.cshtml
@@ -78,7 +78,7 @@ Smartphone Compatible web template, free webdesigns for Nokia, Samsung, LG, Sony
                                     <ul class="nav navbar-nav">
                                         <li><a href="~/Home/Index" class="hvr-underline-from-left"><i class="menu"></i>Home</a></li>
                                         <li><a href="~/Home/About" class="hvr-underline-from-left"><i class="menu"></i>Sobre Nosotros</a></li>
-                                        <li><a href="~/Home/Facilities" class="hvr-underline-from-left"><i class="menu"></i>Facilidades</a></li>
+                                        <li><a href="~/Facilitie/Facilities" class="hvr-underline-from-left"><i class="menu"></i>Facilidades</a></li>
                                         <li><a href="~/Home/HowToGet" class="hvr-underline-from-left"><i class="menu"></i>¿Como Llegar?</a></li>
                                         <li><a href="~/Home/Rates" class="hvr-underline-from-left"><i class="menu"></i>Tarifas</a></li>
                                         <li><a href="~/Reservation/OnlineBook" class="hvr-underline-from-left"><i class="menu"></i>Reservar en Linea</a></li>
